offer-builder/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.1
python-docx==1.1.2
python-multipart==0.0.18
beautifulsoup4==4.12.3
requests==2.32.3
Pillow==11.0.0
google-generativeai==0.8.3
python-dotenv==1.0.1
pydantic==2.10.3
playwright==1.49.0
PyPDF2>=3.0.0
pdfplumber>=0.10.0
